+++ src/SecucardConnect/SecucardConnect.php
@@ -33,7 +33,7 @@ final class SecucardConnect
     /**
      * SDK version
      */
-    const VERSION = '1.19.0';
+    const VERSION = '1.21.0';
 
     /**
      * @var OAuthProvider